requirements.txt
streamlit>=1.31.0
pandas>=2.1.0
requests>=2.31.0
plotly>=5.18.0
anthropic>=0.40.0
yfinance>=0.2.36
openpyxl>=3.1.0
lxml>=5.0.0
google-cloud-storage>=2.14.0
google-cloud-secret-manager>=2.20.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
